move fixtures from conftest.py to separate folders remove conftest.py
